chore: fix issues and refactor
