+++ rest/src/main/java/com/student/rest/service/course/imp/CourseServiceImp.java
@@ -1,6 +1,6 @@
 package com.student.rest.service.course.imp;
 
-import com.student.rest.exceptions.CourseDoesNotExistsException;
+import com.student.rest.exceptions.course.CourseDoesNotExistsException;
 import com.student.rest.exceptions.course.CourseAlreadyExistsException;
 import com.student.rest.model.Course;
 import com.student.rest.repo.CourseRepo;
